+++ build-docker/stack.yml
@@ -14,7 +14,7 @@ services:
     image: solopcloud/adempiere-backend
     restart: always
     environment:
-      - SERVICES_ENABLED=access; business; core; dashboarding; dictionary; enrollment; general_ledger; log; material_management; ui; workflow; store; pos; business_partner; in_out; invoice; order; payment; payroll_action_notice; product; time_control;
+      - SERVICES_ENABLED=access; business; business_partner; core; dashboarding; dictionary; enrollment; file_management; general_ledger; in_out; invoice; log; material_management; order; payment; payroll_action_notice; pos; product; store; time_control; ui; workflow;
       - DB_HOST=database
     ports:
       - 50059:50059
