+++ src/main/java/org/spin/base/util/Services.java
@@ -20,27 +20,28 @@ package org.spin.base.util;
  */
 public enum Services {
     ACCESS("access"),
-    ENROLLMENT("enrollment"),
-    DICTIONARY("dictionary"),
     BUSINESS("business"),
+	BUSINESS_PARTNER("business_partner"),
     CORE("core"),
-	MATERIAL_MANAGEMENT("material_management"),
-    UI("ui"),
     DASHBOARDING("dashboarding"),
+	DICTIONARY("dictionary"),
+	ENROLLMENT("enrollment"),
+	EXTENSION("extension"),
+	FILE_MANAGEMENT("file_management"),
 	GENERAL_LEDGER("general_ledger"),
-    LOG("log"),
-    STORE("store"),
-    POS("pos"),
-    UPDATER("updater"),
-    EXTENSION("extension"),
-    BUSINESS_PARTNER("business_partner"),
-    IN_OUT("in_out"),
-    INVOICE("invoice"),
+	IN_OUT("in_out"),
+	INVOICE("invoice"),
+	LOG("log"),
+	MATERIAL_MANAGEMENT("material_management"),
     ORDER("order"),
     PAYMENT("payment"),
     PAYROLL_ACTION_NOTICE("payroll_action_notice"),
+	POS("pos"),
 	PRODUCT("product"),
+	STORE("store"),
 	TIME_CONTROL("time_control"),
+	UI("ui"),
+	UPDATER("updater"),
 	WORKFLOW("workflow");
 
 	/**	Service Name	*/
