+++ src/main/proto/point_of_sales.proto
@@ -17,11 +17,11 @@ syntax = "proto3";
 option java_multiple_files = true;
 option java_package = "org.spin.backend.grpc.pos";
 option java_outer_classname = "ADempierePOS";
-option objc_class_prefix = "HLW";
 
 import "core_functionality.proto";
 import "base_data_type.proto";
 import "client.proto";
+import "file_management.proto";
 import "time_control.proto";
 
 package data;
@@ -1358,7 +1358,7 @@ message Key {
 	int32 span_y = 9;
 	string product_value = 10;
 	Decimal quantity = 11;
-	ResourceReference resource_reference = 12;
+	file_management.ResourceReference resource_reference = 12;
 }
 
 //	Stock
